requirements.txt
attrs==22.1.0
branca==0.6.0
certifi==2022.9.24
charset-normalizer==2.1.1
contextlib2==21.6.0
coverage==6.5.0
exceptiongroup==1.0.1
fit2gpx==0.0.6
fitdecode==0.10.0
folium==0.13.0
gpxpy==1.5.0
idna==3.4
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
schema==0.7.5
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
